Moving response types to separate files
